Sets a global variable for column width in queues.py
Columns will soon need an increase in width due to the job number
rolling over to 10^6. This commit sets a global variable (I know those
are bad, but the y make life easy) for the width of a COLUMN.
Eventually, all these options should be passed as a dictionary to the
print function.